+++ package.json
@@ -56,7 +56,7 @@
     "prettier": "^3.6.2",
     "release-it": "^19.0.3",
     "tsx": "^4.19.0",
-    "typescript": "^5.4.5",
+    "typescript": "^5.8.3",
     "vitest": "^3.2.4",
     "@restatedev/restate-sdk-testcontainers": "^1.7.0"
   },
