+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const RESTATE_PROMISE_RESOLVE = "restate.promise.resolve";
+export const RESTATE_PROMISE_REJECT = "restate.promise.reject";
